+++ Client/World.cs
@@ -34,7 +34,7 @@ namespace Client
             }
         }
 
-        public override void Tick(double delta)
+        public override void Tick(float delta)
         {
 
         }
