+++ Server/World.cs
@@ -41,7 +41,7 @@ namespace Server
             }
         }
 
-        public override void Tick(double delta)
+        public override void Tick(float delta)
         {
 
         }
